Refactor MainActivity to a role hub, and move entrant demo to EntrantActivity
