+++ app/src/main/java/com/example/impact/view/AdminActivity.java
@@ -1,4 +1,15 @@
 package com.example.impact.view;
 
-public class AdminActivity {
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+import com.example.impact.R;
+
+public class AdminActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_admin);
+    }
 }
